Moved extension to own file and improved variable names
