+++ Mooov/Mooov/Controllers/LogInViewController.swift
@@ -15,14 +15,12 @@ class LogInViewController: UIViewController {
     
     @IBAction private func loginButtonPressed(_ sender: Any) {
         if usernameField.text == "Admin" && passwordField.text == "TestPass123"{
-            let vc = storyboard?.instantiateViewController(identifier: "landingPageVC") as? LandingPageViewController
-            
-            guard let vc = vc else {
+            guard let landingPageScreen = storyboard?.instantiateViewController(identifier: "landingPageVC") as? LandingPageViewController else {
                 return
             }
             
-            vc.modalPresentationStyle = .fullScreen
-            present(vc, animated: true, completion: nil)
+            landingPageScreen.modalPresentationStyle = .fullScreen
+            present(landingPageScreen, animated: true, completion: nil)
             
         } else {
             self.displayAlert(title: "Log in failed",
@@ -33,13 +31,5 @@ class LogInViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-    }
-}
-
-extension UIViewController {
-    func displayAlert(title: String, message: String, buttonTitle: String) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: UIAlertController.Style.alert)
-        alert.addAction(UIAlertAction(title: buttonTitle, style: UIAlertAction.Style.default, handler: nil))
-        self.present(alert, animated: true, completion: nil)
     }
 }
